+++ flutter_local_notifications/android/src/main/java/com/dexterous/flutterlocalnotifications/ScheduledNotificationReceiver.java
@@ -22,31 +22,16 @@ public class ScheduledNotificationReceiver extends BroadcastReceiver {
 
     @Override
     public void onReceive(final Context context, Intent intent) {
-        NotificationManagerCompat notificationManager = NotificationManagerCompat.from(context);
         String notificationDetailsJson = intent.getStringExtra(FlutterLocalNotificationsPlugin.NOTIFICATION_DETAILS);
-        boolean repeat = intent.getBooleanExtra(FlutterLocalNotificationsPlugin.REPEAT, false);
-
-        // TODO: remove this branching logic as it's legacy code to fix an issue where notifications weren't reporting the correct time
-        if(StringUtils.isNullOrEmpty(notificationDetailsJson)) {
-            Notification notification = intent.getParcelableExtra(FlutterLocalNotificationsPlugin.NOTIFICATION);
-            notification.when = System.currentTimeMillis();
-            int notificationId = intent.getIntExtra(FlutterLocalNotificationsPlugin.NOTIFICATION_ID,
-                    0);
-            notificationManager.notify(notificationId, notification);
-            if (repeat) {
-                return;
-            }
-            FlutterLocalNotificationsPlugin.removeNotificationFromCache(notificationId, context);
-        } else {
-            Gson gson = FlutterLocalNotificationsPlugin.buildGson();
-            Type type = new TypeToken<NotificationDetails>() {
-            }.getType();
-            NotificationDetails notificationDetails  = gson.fromJson(notificationDetailsJson, type);
-            FlutterLocalNotificationsPlugin.showNotification(context, notificationDetails);
-            if (repeat) {
-                return;
-            }
+        Gson gson = FlutterLocalNotificationsPlugin.buildGson();
+        Type type = new TypeToken<NotificationDetails>() {
+        }.getType();
+        NotificationDetails notificationDetails  = gson.fromJson(notificationDetailsJson, type);
+        FlutterLocalNotificationsPlugin.showNotification(context, notificationDetails);
+        if(notificationDetails.scheduledNotificationRepeatFrequency == null && notificationDetails.repeatInterval == null) {
             FlutterLocalNotificationsPlugin.removeNotificationFromCache(notificationDetails.id, context);
+        } else {
+            FlutterLocalNotificationsPlugin.scheduleNextNotification(context, notificationDetails);
         }
 
     }