+++ flutter_local_notifications/android/src/main/java/com/dexterous/flutterlocalnotifications/models/NotificationDetails.java
@@ -105,6 +105,7 @@ public class NotificationDetails {
 
     private static final String SCHEDULED_DATE_TIME = "scheduledDateTime";
     private static final String TIMEZONE_NAME = "timezoneName";
+    private static final String SCHEDULED_NOTIFICATION_REPEAT_FREQUENCY = "scheduledNotificationRepeatFrequency";
 
 
     public Integer id;
@@ -156,6 +157,7 @@ public class NotificationDetails {
     public Boolean showWhen;
     public String scheduledDateTime;
     public String timezoneName;
+    public ScheduledNotificationRepeatFrequency scheduledNotificationRepeatFrequency;
     public Long when;
 
 
@@ -171,6 +173,9 @@ public class NotificationDetails {
         notificationDetails.body = (String) arguments.get(BODY);
         notificationDetails.scheduledDateTime = (String) arguments.get(SCHEDULED_DATE_TIME);
         notificationDetails.timezoneName = (String) arguments.get(TIMEZONE_NAME);
+        if(arguments.containsKey(SCHEDULED_NOTIFICATION_REPEAT_FREQUENCY)) {
+            notificationDetails.scheduledNotificationRepeatFrequency = ScheduledNotificationRepeatFrequency.values()[(Integer) arguments.get(SCHEDULED_NOTIFICATION_REPEAT_FREQUENCY)];
+        }
         if (arguments.containsKey(MILLISECONDS_SINCE_EPOCH)) {
             notificationDetails.millisecondsSinceEpoch = (Long) arguments.get(MILLISECONDS_SINCE_EPOCH);
         }