+++ flutter_local_notifications/lib/src/types.dart
@@ -53,3 +53,8 @@ class Time {
     };
   }
 }
+
+enum ScheduledNotificationRepeatFrequency {
+  Daily,
+  Weekly,
+}